Refactor scriptbuilder
